requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
FastAPI-SQLAlchemy==0.2.1
pydantic==2.3.0
alembic==1.12.0
psycopg2==2.9.7
psycopg-binary==3.1.10
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
